requirements.txt
bunch==1.0.1
dm_control
dm_env==1.6
einops==0.7.0
glfw==2.6.5
gym==0.23.0
haven_ai==0.7.3
hydra-core==1.1.0
imageio==2.9.0
lxml==5.1.0
matplotlib==3.9.0
memory_profiler==0.61.0
mujoco_py==2.1.2.14
numpy==1.24.4
omegaconf==2.1.2
open_clip_torch==2.24.0
opencv_python==4.9.0.80
pandas==2.2.2
pytest==8.0.2
robosuite==1.4.1
scipy==1.9.3
setuptools==59.5.0
tensorboard
tensorboard_data_server
tensorflow==2.15.0.post1
termcolor==1.1.0
torch==2.2.0 
torchvision 
tqdm==4.66.1
wandb==0.16.4
flash_attn==2.5.7
peft==0.4.0
transformers==4.31.0
hydra-submitit-launcher==1.1.5
moviepy==1.0.3
imageio[ffmpeg]
# for the demo: gradio==4.36.1